move setMainClockCountdown to listeventlistener
